% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr.R
\name{aggregate_dmr_landings_county}
\alias{aggregate_dmr_landings_county}
\title{Aggregate by DMR landings data by county, year and species}
\usage{
aggregate_dmr_landings_county(
  x = read_dmr_landings("modern"),
  collapse = "none"
)
}
\arguments{
\item{x}{DMR landings data}

\item{collapse}{chr, "none" (default) or "year" to aggregate years etc}
}
\value{
summary table
}
\description{
Aggregate by DMR landings data by county, year and species
}
