% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_nao}
\alias{fetch_nao}
\title{Fetch NAO}
\usage{
fetch_nao(
  url =
    "https://www.cpc.ncep.noaa.gov/products/precip/CWlink/pna/norm.nao.monthly.b5001.current.ascii.table",
  path = oame_path("NAO")
)
}
\arguments{
\item{url}{chr, the URL for of the data source}

\item{path}{chr, the destination path}
}
\value{
data frame of most recent data
}
\description{
Fetch NAO
}
