% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdat.R
\name{map_hurdat}
\alias{map_hurdat}
\title{Build a map of the hurricane paths faceted by user-specified epoch}
\usage{
map_hurdat(
  x = read_hurdat(),
  color_by = c("wind_max_sus", "duration", "min_pres")[1],
  epoch = c(10, 25, 50, 100)[2],
  counties = sf::st_crop(sf::st_geometry(read_me_counties()), x),
  coast = read_coast()
)
}
\arguments{
\item{x}{HURDAT sf object}

\item{color_by}{chr, one of "wind_sus_max", "duration", or "min_pres"}

\item{epoch}{num, the number of consecutive years to group by}

\item{counties}{spatial geometry of counties (maine)}

\item{coast}{spatial geometry of coastline (gulf of maine)}
}
\value{
ggplot2 object
}
\description{
Build a map of the hurricane paths faceted by user-specified epoch
}
