% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs.R
\name{map_labs}
\alias{map_labs}
\title{Place ME research institutions on an interactive leaflet map}
\usage{
map_labs(x)
}
\arguments{
\item{x}{tibble of institutions with lat/lon locations}
}
\value{
a leaflet map
}
\description{
Place ME research institutions on an interactive leaflet map
}
