% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oame-package.R
\docType{package}
\name{oame-package}
\alias{oame-package}
\alias{oame}
\title{Ocean Accounts Maine}
\description{
Serve Maine Ocean Accounts data.
}
\author{
\strong{Maintainer}: Ben Tupper \email{btupper@bigelow.org}

Authors:
\itemize{
  \item Ben Tupper \email{btupper@bigelow.org}
  \item Johnathan Evanilla \email{jevanilla@bigelow.org}
}

}
\keyword{internal}
