% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{oame_path}
\alias{oame_path}
\title{Retrieve and/or build a data path}
\usage{
oame_path(..., root = root_data_path())
}
\arguments{
\item{...}{chr, file path segments ala \code{\link[base:file.path]{base::file.path()}}. These must come before
the \code{root} argument.}

\item{root}{chr the root data path}
}
\value{
a complete path specification
}
\description{
Retrieve and/or build a data path
}
