% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{plot_census}
\alias{plot_census}
\title{Plots county level census data for Maine coastal counties}
\usage{
plot_census(plot_data = oame::read_census(), y_var = "population")
}
\arguments{
\item{plot_data}{tibble of filtered data for a specific region of Maine, or the whole state}

\item{y_var}{character string variable for y axis}
}
\value{
a ggplot
}
\description{
Plots county level census data for Maine coastal counties
}
