% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr.R
\name{plot_landings}
\alias{plot_landings}
\title{Plots DMR landings data over time for different variables}
\usage{
plot_landings(x, spp, varname)
}
\arguments{
\item{x}{tibble of landings data from DMR}

\item{varname}{character string variable for y axis}
}
\value{
a ggplot
}
\description{
Plots DMR landings data over time for different variables
}
