% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_monthly_timeseries}
\alias{plot_monthly_timeseries}
\alias{plot_monthly_climatology}
\alias{plot_nao}
\alias{plot_amo}
\title{Plot a timeseries}
\usage{
plot_monthly_timeseries(x = read_nao(), title = "NAO", facet = FALSE)

plot_monthly_climatology(x = read_nao(), title = "NAO")

plot_nao(x = read_nao(), type = c("timeseries", "monthly", "climatology")[1])

plot_amo(x = read_amo(), type = c("timeseries", "monthly", "climatology")[1])
}
\arguments{
\item{x}{data frame with columns "date" and "value"}

\item{title}{chr, the title for the plot}

\item{facet}{logical, if TRUE facet by month}

\item{smooth}{chr or NULL, if not NULL then the smoothing type to use
otherwise don't apply smoothing}
}
\value{
ggplot object
}
\description{
Plot a timeseries
}
