% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean-econ.R
\name{plot_ocean_economy}
\alias{plot_ocean_economy}
\title{Plots ocean economy sectors over time for different variables}
\usage{
plot_ocean_economy(plot_data, y_var = "gdp")
}
\arguments{
\item{plot_data}{tibble of filtered data for a specific region of Maine, or the whole state}

\item{y_var}{character string variable for y axis}
}
\value{
a ggplot
}
\description{
Plots ocean economy sectors over time for different variables
}
