% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa-tides.R
\name{plot_tide_climatology}
\alias{plot_tide_climatology}
\alias{plot_tide}
\title{Plot monthly tide and tide climatology}
\usage{
plot_tide_climatology(x = read_tide())

plot_tide(x = read_tide(), type = c("timeseries", "monthly", "climatology")[1])
}
\arguments{
\item{x}{data frame of tide data}

\item{type}{chr one "timeseries" (default), "monthly" or "climatology"}
}
\value{
station faceted ggplot object
}
\description{
Plot monthly tide and tide climatology
}
