% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr.R
\name{plotly_species_by_county}
\alias{plotly_species_by_county}
\title{Make a chloropleth map by county for one or more species and
and one or more years.}
\usage{
plotly_species_by_county(
  x = read_dmr_landings("modern"),
  spp = "Clam Soft",
  years = "recent",
  varname = "trip_n",
  counties = read_me_counties(crs = 3857),
  style = c("plain", "cartogram")[1]
)
}
\arguments{
\item{x}{table of DMR landings}

\item{spp}{chr, one or more species}

\item{years}{chr or num, "recent" for the most recent year for the species,
or specify your own year.  If you specify multiple years the data is
aggregated.}

\item{varname}{chr the name of the variable to map}

\item{counties}{sf table with "county" attribute and in 3857 CRS}

\item{style}{chr, one of "plain" or "cartogram"}
}
\value{
map plot object
}
\description{
Make a chloropleth map by county for one or more species and
and one or more years.
}
