% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr.R
\name{prep_dmr_landings_county}
\alias{prep_dmr_landings_county}
\title{Prep landings data form county analyses}
\usage{
prep_dmr_landings_county(x = read_dmr_landings("modern"))
}
\arguments{
\item{x}{DMR landings table}
}
\value{
reduced DMR table to have valid county names
}
\description{
Prep landings data form county analyses
}
