% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_census}
\alias{read_census}
\title{Read Maine census data}
\usage{
read_census(which = c("county", "county_subdivision")[1])
}
\arguments{
\item{which}{string choice of geometry level for census data}
}
\value{
tibble
}
\description{
Read Maine census data
}
