% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_coast}
\alias{read_coast}
\title{Read coastline geometry}
\usage{
read_coast(filename = oame_path("NaturalEarth/gom_coast.rds"))
}
\arguments{
\item{filename}{chr, the path specification for the file}
}
\value{
sfc geometry for the coast
}
\description{
Read coastline geometry
}
