% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read the configuration file}
\usage{
read_config(filename = "~/.oame")
}
\arguments{
\item{filename}{chr, the path specification for the config file}
}
\value{
a configuration list
}
\description{
Read the configuration file
}
