% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_hab_index}
\alias{read_hab_index}
\title{Reads HAB index}
\usage{
read_hab_index(filename = oame_path("hab", "hab-index-2026-08-11.csv.gz"))
}
\arguments{
\item{filename}{chr, the path specification for the file}
}
\value{
tibble
}
\description{
Reads HAB index
}
