% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_hurdat}
\alias{read_hurdat}
\title{Read hurricane data}
\usage{
read_hurdat(filename = oame_path("NOAA/hurdat2/gom_storms.rds"))
}
\arguments{
\item{filename}{chr, the path specification for the file}
}
\value{
SF LINESTRING data by storm
}
\description{
Read hurricane data
}
