% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_labs}
\alias{read_labs}
\title{Reads research location data}
\usage{
read_labs(filename = oame_path("LABS", "me_marine_labs.csv.gz"))
}
\arguments{
\item{filename}{chr, the path specification for the file}
}
\value{
tibble
}
\description{
Reads research location data
}
