% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_me_coastal_counties}
\alias{read_me_coastal_counties}
\title{Read Maine coastal county FIPS codes (geoid)}
\usage{
read_me_coastal_counties(
  filename = oame_path("ME", "me_coastal_county_geoid.csv.gz")
)
}
\arguments{
\item{filename}{path specification for the file}
}
\value{
tibble
}
\description{
Read Maine coastal county FIPS codes (geoid)
}
