% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_monthly_fwf}
\alias{read_monthly_fwf}
\alias{read_nao}
\alias{read_amo}
\title{Read a fixed-width format monthly file ala AMO and NAO}
\usage{
read_monthly_fwf(
  filename = oame_path("AMO", "amo.csv"),
  na = c("", "NA", "NaN", "-99.990"),
  form = c("wide", "long")[1]
)

read_nao(filename = oame_path("NAO", "nao.csv"), form = c("wide", "long")[2])

read_amo(filename = oame_path("AMO", "amo.csv"), form = c("wide", "long")[2])
}
\arguments{
\item{filename}{chr the path specification for the file}

\item{na}{chr, one or more flags indicating no data}

\item{form}{chr, either "wide" (default) or "long"}
}
\value{
tibble
}
\description{
\code{read_monthly_fwf} reads a fixed width text file with unlabeled
columns year, Jan, Feb, ..., Nov, Dec.  \code{read_amo} and \code{read_nao}
are wrappers around \code{read_monthly_fwf} that preferentially return the
long format of data.
}
\details{
'
}
