% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nao.R
\name{read_nao}
\alias{read_nao}
\title{Read NAO data}
\usage{
read_nao(filename = oame_path("NAO", "nao.csv"))
}
\arguments{
\item{filename}{chr the path specification for the file}

\item{form}{chr one of "long" or "wide" (default)}
}
\value{
tibble
}
\description{
Read NAO data
}
