% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa-tides.R
\name{read_tide}
\alias{read_tide}
\title{Read tide data}
\usage{
read_tide(
  x = tide_stations(),
  path = oame_path("NOAA", "tides"),
  drop = c("Cutler", "none")[1]
)
}
\arguments{
\item{x}{data frame of tide stations identifiers}

\item{path}{chr, the root data path}

\item{drop}{chr, the names of stations to drop, "none" to drop none.}
}
\value{
data frame of tide station data
}
\description{
Read tide data
}
