% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{root_data_path}
\alias{root_data_path}
\title{Retrieve the root data path}
\usage{
root_data_path()
}
\arguments{
\item{filename}{chr the name where the root data path is stored (as a yaml)}
}
\value{
file path to the data directory
}
\description{
Retrieve the root data path
}
