% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa-tides.R
\name{tide_stations}
\alias{tide_stations}
\title{Retrieve a table of select stations}
\usage{
tide_stations(root = "https://tidesandcurrents.noaa.gov/sltrends/data")
}
\arguments{
\item{root}{chr, the root URL to access data}
}
\value{
data frame with "name", "id" and "url"
}
\description{
Retrieve a table of select stations
}
