% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa-tides.R
\name{fetch_tides}
\alias{fetch_tides}
\title{Fetch tide data}
\usage{
fetch_tides(x = tide_stations(), path = oame_path("NOAA", "tides"))
}
\arguments{
\item{x}{data frame of tide stations}

\item{path}{chr, the root data path}
}
\value{
the input with an added "filename" column
}
\description{
Fetch tide data
}
