% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_me_counties}
\alias{read_me_counties}
\title{Read Maine county boundary files}
\usage{
read_me_counties(
  filename = oame_path("ME", "Maine_County_Boundary_Polygons_Dissolved_Feature.gpkg"),
  simplify = TRUE,
  crs = NULL
)
}
\arguments{
\item{filename}{path specification for the file}

\item{simplify}{logical, if TRUE simplify to just "county" from the complete
table (not the case change from "COUNTY")}

\item{crs}{NULL or a crs to transform the geometry}
}
\value{
sf MULTIPLOYGON table
}
\description{
Read Maine county boundary files
}
