% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_me_counties}
\alias{read_me_counties}
\title{Read Maine county boundary files

@export
@param filename path specification for the file
@param simplify logical, if TRUE simplify to just "county" from the complete
table (not the case change from "COUNTY")
@param crs NULL or a crs to transform the geometry
@return sf MULTIPLOYGON table}
\usage{
read_me_counties(
  filename = oame_path("ME", "Maine_County_Boundary_Polygons_Dissolved_Feature.geojson"),
  simplify = TRUE,
  crs = NULL
)
}
\description{
Read Maine county boundary files

@export
@param filename path specification for the file
@param simplify logical, if TRUE simplify to just "county" from the complete
table (not the case change from "COUNTY")
@param crs NULL or a crs to transform the geometry
@return sf MULTIPLOYGON table
}
